+++ packages/core/src/Engine.ts
@@ -1,15 +1,16 @@
 import { BoundingBox, Vector3 } from "@oasis-engine/math";
 import { Color } from "@oasis-engine/math/src/Color";
-import { ColorSpace, ShadowCascadesMode, ShadowMode, ShadowResolution } from ".";
 import { ResourceManager } from "./asset/ResourceManager";
 import { Event, EventDispatcher, Logger, Time } from "./base";
+import { GLCapabilityType } from "./base/Constant";
 import { Canvas } from "./Canvas";
 import { ComponentsManager } from "./ComponentsManager";
 import { EngineSettings } from "./EngineSettings";
 import { Entity } from "./Entity";
+import { ColorSpace } from "./enums/ColorSpace";
 import { InputManager } from "./input";
 import { LightManager } from "./lighting/LightManager";
-import { Material, RenderQueueType } from "./material";
+import { Material } from "./material/Material";
 import { PhysicsManager } from "./physics";
 import { IHardwareRenderer } from "./renderingHardwareInterface";
 import { ClassPool } from "./RenderPipeline/ClassPool";
@@ -21,15 +22,23 @@ import { SpriteMaskManager } from "./RenderPipeline/SpriteMaskManager";
 import { TextRenderElement } from "./RenderPipeline/TextRenderElement";
 import { Scene } from "./Scene";
 import { SceneManager } from "./SceneManager";
-import { BlendFactor, BlendOperation, ColorWriteMask, CompareFunction, CullMode, Shader } from "./shader";
+import { BlendFactor } from "./shader/enums/BlendFactor";
+import { BlendOperation } from "./shader/enums/BlendOperation";
+import { ColorWriteMask } from "./shader/enums/ColorWriteMask";
+import { CompareFunction } from "./shader/enums/CompareFunction";
+import { CullMode } from "./shader/enums/CullMode";
+import { RenderQueueType } from "./shader/enums/RenderQueueType";
+import { Shader } from "./shader/Shader";
 import { ShaderMacro } from "./shader/ShaderMacro";
 import { ShaderMacroCollection } from "./shader/ShaderMacroCollection";
 import { ShaderPass } from "./shader/ShaderPass";
 import { ShaderPool } from "./shader/ShaderPool";
 import { ShaderProgramPool } from "./shader/ShaderProgramPool";
 import { RenderState } from "./shader/state/RenderState";
+import { ShadowCascadesMode } from "./shadow/enum/ShadowCascadesMode";
+import { ShadowMode } from "./shadow/enum/ShadowMode";
+import { ShadowResolution } from "./shadow/enum/ShadowResolution";
 import { Texture2D, Texture2DArray, TextureCube, TextureCubeFace, TextureFormat } from "./texture";
-import { GLCapabilityType } from "./base/Constant";
 
 ShaderPool.init();
