+++ packages/core/src/RenderPipeline/MeshRenderElement.ts
@@ -2,6 +2,8 @@ import { Mesh } from "../graphic/Mesh";
 import { SubMesh } from "../graphic/SubMesh";
 import { Material } from "../material/Material";
 import { Renderer } from "../Renderer";
+import { ShaderPass } from "../shader/ShaderPass";
+import { RenderState } from "../shader/state/RenderState";
 import { RenderElement } from "./RenderElement";
 
 /**
@@ -13,10 +15,19 @@ export class MeshRenderElement extends RenderElement {
   /** Sub mesh. */
   subMesh: SubMesh;
 
-  setValue(component: Renderer, mesh: Mesh, subMesh: SubMesh, material: Material): void {
+  setValue(
+    component: Renderer,
+    mesh: Mesh,
+    subMesh: SubMesh,
+    material: Material,
+    renderState: RenderState,
+    shaderPass: ShaderPass
+  ): void {
     this.component = component;
     this.mesh = mesh;
     this.subMesh = subMesh;
     this.material = material;
+    this.renderState = renderState;
+    this.shaderPass = shaderPass;
   }
 }
